pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "astra-forge"
version = "2.0.0"
description = "Agent SDK orchestrator for the Astra development lifecycle plugin"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Prateek Jain" },
]
dependencies = [
    "claude-agent-sdk>=0.1.50",
    "pydantic>=2.0",
    "rich>=13.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[project.scripts]
astra-forge = "orchestrator.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["orchestrator"]

[tool.ruff]
target-version = "py310"
line-length = 100
